==> Data Cleaning and Fixing/Combine 11v11 and Prem/step1-pre-merge-clean.ipynb <==
import pandas as pd
import time

elevenData = pd.read_csv("11v11_data.csv")
premData = pd.read_csv("prem_data.csv")

print("Size of 11v11 Dataset: {} games".format(len(elevenData)))
print("Size of Prem Dataset: {} games".format(len(premData)))
print("Target size: 10000 total games")

s = elevenData["Date"].tolist()
#for date in s:
    #if len(str(date)) != 8:
        #print(date)

early_dict = {"1992":0,"1993":1,"1994":2}
n = 3
for i in range(1995,2018):
    early_dict[str(i)] = n
    n += 1
early_dict["2018"] = 26

def get_season_from_date(row):
    d = str(row['Date'])
    date = d
    if len(d) != 8:
        date = "0" + d
    newdate1 = time.strptime(date, "%d%m%Y")
    newdate2 = time.strptime("3105"+date[-4:], "%d%m%Y")
    early = newdate1 < newdate2
    
    
    season = early_dict.get(str(date[-4:]))
    if not(early):
        season += 1
    return int(season)

elevenData["Season"] = elevenData.apply(get_season_from_date,axis=1)

prev = 0
season_col = elevenData["Season"].tolist()
for i in range(len(season_col)-1):
    cur = int(season_col[i])
    if prev != cur:
        print(cur,end=",")
    prev = cur

def convert_date_to_int(row):
    return int(str(row["Date"]))
premData["Date"] = premData.apply(convert_date_to_int,axis=1)

premData["Season"] = premData.apply(get_season_from_date,axis=1)

prev = 0
season_col = premData["Season"].tolist()
for i in range(len(season_col)-1):
    cur = int(season_col[i])
    if prev != cur:
        print(cur,end=",")
    prev = cur

premData.head()

premData.drop(['Stadium'],axis=1,inplace=True)
premData.head()

elevenData.head()

def fix_date_column(row):
    raw_date = str(row['Date'])
    date = raw_date
    if len(raw_date) != 8:
        date = "0" + raw_date
    if len(date) != 8:
        print(date)
        return date
    else:
        day = str(date[0:2])
        month = str(date[2:4])
        year = str(date[4:8])
        final_date = str(month) + "-" + str(day) + "-" + str(year)
        return final_date
    
premData["Date"] = premData.apply(fix_date_column,axis=1)
elevenData["Date"] = elevenData.apply(fix_date_column,axis=1)

premData.head()

elevenData.head()

premData["Date"] = pd.to_datetime(premData.Date)
elevenData["Date"] = pd.to_datetime(elevenData.Date)

premData.head()

premData.sort_values(by=['Date','Home Team'],inplace=True)
elevenData.sort_values(by=['Date','Home Team'],inplace=True)

premData.head(100)

premData.to_csv("prem_cleaned1.csv",index=False)
elevenData.to_csv("11v11_cleaned1.csv",index=False)
