==> Data Cleaning and Fixing/Combine 11v11 and Prem/step2-remove-non-pl-games.ipynb <==
import pandas as pd
import time
import numpy as np

all_teams_seasons_dict = {}
all_teams_seasons_dict['arsenal'] = [i for i in range(1,27)]
all_teams_seasons_dict['aston villa'] = [i for i in range(1,25)]
all_teams_seasons_dict['barnsley'] = [6]
all_teams_seasons_dict['birmingham'] = [11,12,13,14,16,18,19]
all_teams_seasons_dict['blackburn'] = [1,2,3,4,5,6,7,10] + [i for i in range(11,21)]
all_teams_seasons_dict['blackpool'] = [19]
all_teams_seasons_dict['bolton'] = [4,6] + [i for i in range(10,21)]
all_teams_seasons_dict['bournemouth'] = [i for i in range(24,27)] 
all_teams_seasons_dict['bradford'] = [8,9]
all_teams_seasons_dict['brighton'] = [26]
all_teams_seasons_dict['burnley'] = [18,23,25,26]
all_teams_seasons_dict['cardiff'] = [22]
all_teams_seasons_dict['charlton'] = [7,9,10,11,12,13,14,15]
all_teams_seasons_dict['chelsea'] = [i for i in range(1,27)]
all_teams_seasons_dict['coventry'] = [i for i in range(1,10)]
all_teams_seasons_dict['crystal palace'] = [1,3,6,13,22,23,24,25,26]
all_teams_seasons_dict['derby'] = [5,6,7,8,9,10,16]
all_teams_seasons_dict['everton'] = [i for i in range(1,27)]
all_teams_seasons_dict['fulham'] = [i for i in range(10,23)]
all_teams_seasons_dict['huddersfield'] = [26]
all_teams_seasons_dict['hull'] = [17,18,19,22,23,24,25]
all_teams_seasons_dict['ipswich'] = [1,2,3,9,10]
all_teams_seasons_dict['leeds'] = [i for i in range(1,13)]
all_teams_seasons_dict['leicester'] = [3,5,6,7,8,9,10,12,23,24,25,26]
all_teams_seasons_dict['liverpool'] = [i for i in range(1,27)]
all_teams_seasons_dict['man city'] = [1,2,3,4,9] + [i for i in range(11,27)]
all_teams_seasons_dict['man utd'] = [i for i in range(1,27)]
all_teams_seasons_dict['middlesbrough'] = [1,4,5,7,8,9,10,11,12,13,14,15,16,17,25]
all_teams_seasons_dict['newcastle'] = [i for i in range(2,18)] + [19,20,21,22,23,24,26]
all_teams_seasons_dict['norwich'] = [1,2,3,13,20,21,22,24]
all_teams_seasons_dict['nottingham forest'] = [1,3,4,5,7]
all_teams_seasons_dict['oldham'] = [1,2]
all_teams_seasons_dict['portsmouth'] = [12,13,14,15,16,17,18]
all_teams_seasons_dict['qpr'] = [1,2,3,4,20,21,23]
all_teams_seasons_dict['reading'] = [15,16,21]
all_teams_seasons_dict['sheffield utd'] = [1,2,15]
all_teams_seasons_dict['sheffield wed'] = [1,2,3,4,5,6,7,8]
all_teams_seasons_dict['southampton'] = [1,2,3,4,5,6,7,8,9,10,11,12,13,21,22,23,24,25,26]
all_teams_seasons_dict['spurs'] = [i for i in range(1,27)]
all_teams_seasons_dict['stoke'] = [i for i in range(17,27)]
all_teams_seasons_dict['sunderland'] = [5,8,9,10,11,14,16,17,18,19,20,21,22,23,24,25]
all_teams_seasons_dict['swansea'] = [20,21,22,23,24,25,26]
all_teams_seasons_dict['swindon'] = [2]
all_teams_seasons_dict['watford'] = [8,15,24,25,26]
all_teams_seasons_dict['west brom'] = [11,13,14,17,19,20,21,22,23,24,25,26]
all_teams_seasons_dict['west ham'] = [2,3,4,5,6,7,8,9,10,11,14,15,16,17,18,19,21,22,23,24,25,26]
all_teams_seasons_dict['wigan'] = [14,15,16,17,18,19,20,21]
all_teams_seasons_dict['wimbledon'] = [1,2,3,4,5,6,7,8]
all_teams_seasons_dict['wolves'] = [12,18,19,20]
#all_teams_seasons_dict

prem = pd.read_csv("prem_cleaned1.csv")
eleven = pd.read_csv("11v11_cleaned1.csv")

prem.head()

eleven.head()

def test_prem_game(row):
    home = str(row['Home Team'])
    season = int(row['Season'])
    team_seasons = all_teams_seasons_dict.get(home)
    if len(team_seasons) < 1:
        return False
    if season in team_seasons:
        return True
    else:
        #print("{}: {} v {}".format(row['Date'],home,str(row['Away Team'])))
        return False

prem['isPrem'] = prem.apply(test_prem_game,axis=1)

eleven['isPrem'] = eleven.apply(test_prem_game,axis=1)

premNew = prem[prem.isPrem]
elevenNew = eleven[eleven.isPrem]

print("{} -> {}".format(len(prem),len(premNew)))
print("{} -> {}".format(len(eleven),len(elevenNew)))

premNew['Season'].value_counts()

prem22 = premNew.query("Season == 22")
premNew['hasRef'] = [not(x) for x in premNew['Referee'].isna().tolist()]
premRefsOnly = premNew[premNew.noRef]
premRefsOnly['Season'].value_counts()

premNew2 = premRefsOnly
premNew2.head()

eleven['Season'].value_counts()

eleven_s10 = elevenNew.query("Season == 10")
eleven_s10.to_csv("eleven_s10.csv",index=False)

premNew2.head()

eleven.head()

premNew2.drop(['isPrem','noRef','hasRef'],inplace=True,axis=1)
premNew2.head()

premNew2.to_csv("prem_removed2.csv",index=False)
eleven.to_csv("eleven_removed2.csv",index=False)

premNew3 = pd.read_csv("prem_manually_removed3.csv")
premNew3['Season'].value_counts()
premNew3['hasScore'] = [not(x) for x in premNew3['Score'].isna()]
premNew4 = premNew3[premNew3.hasScore]
premNew4['Season'].value_counts()

premNew4.drop('hasScore',inplace=True,axis=1)

premNew4.head()

premNew4.to_csv("prem_cleaned4.csv",index=False)
