==> Data Cleaning and Fixing/Combine 11v11 and Prem/step3-data-finalise-cleaning.ipynb <==
import pandas as pd
import time
import numpy as np

prem = pd.read_csv("prem_cleaned4.csv")
eleven = pd.read_csv("eleven_removed2.csv")

prem.head()

prem.columns = ['premid','date','home','away','score','ref','season']
prem = prem[['premid','season','date','home','score','away','ref']]
prem.head()

eleven.drop('isPrem',inplace=True,axis=1)
eleven.head()

eleven.columns = ['11id','date','home','away','stadium','ref','h_players','a_players','h_scorers','a_scorers','season']
eleven = eleven[['11id','season','date','home','away','ref','stadium','h_scorers','a_scorers','h_players','a_players']]
eleven.head()

len(prem)

len(eleven)

prem.to_csv("prem_finalised.csv",index=False)
eleven.to_csv("eleven_finalised.csv",index=False)
