==> Data Cleaning and Fixing/Combine 11v11 and Prem/step4-merge-and-clean.ipynb <==
import pandas as pd
import time
import numpy as np

prem = pd.read_csv("prem_finalised.csv")
eleven = pd.read_csv("eleven_finalised.csv")

prem.columns = [str(x)+"2" for x in list(prem.columns.values)]
df = pd.concat([prem,eleven],axis=1)

df.head()

def test_equal(row,col1,col2):
    val1 = str(row[col1])
    val2 = str(row[col2])
    return (val1 == val2)

df['home_equal'] = df.apply(test_equal,args=('home','home2'),axis=1)
df['away_equal'] = df.apply(test_equal,args=('away','away2'),axis=1)
df['refs_equal'] = df.apply(test_equal,args=('ref','ref2'),axis=1)

df['home_equal'].value_counts()

df['away_equal'].value_counts()

df['refs_equal'].value_counts()

teams_not_equal = df.query("home_equal != True | away_equal != True")

teams_not_equal.head()

teams_not_equal.to_csv("teams_not_equal.csv",index=False)

# after some manual fixes

prem_fixed = pd.read_csv("prem_finalised_fix_dates.csv")
eleven_fixed = pd.read_csv("eleven_finalised.csv")
prem_fixed.sort_values(by=['date','home'],axis=0,inplace=True)
prem_fixed.reset_index(drop=True,inplace=True)
eleven_fixed.sort_values(by=['date','home'],axis=0,inplace=True)
eleven_fixed.reset_index(drop=True,inplace=True)

prem_fixed.head()

season_12_prem = prem_fixed.query("season == 12")
season_12_prem.head(100)[30:100]

prem_fixed.columns = [str(x)+"2" for x in list(prem_fixed.columns.values)]
df2 = pd.concat([prem_fixed,eleven_fixed],axis=1)

df2.head()

df2['home_equal'] = df2.apply(test_equal,args=('home','home2'),axis=1)
df2['away_equal'] = df2.apply(test_equal,args=('away','away2'),axis=1)
df2['refs_equal'] = df2.apply(test_equal,args=('ref','ref2'),axis=1)

df2['home_equal'].value_counts()

df2['away_equal'].value_counts()

df2['refs_equal'].value_counts()

df2.to_csv("initial_merge_all_games_matching.csv",index=False)

df_refs_different = df2.query("refs_equal == False")

df_refs_different.shape

df_refs_different.to_csv("refs_different_merged.csv",index=False)
