==> Data Cleaning and Fixing/Combine 11v11 and Prem/step5-fix-mismatching-refs.ipynb <==
import pandas as pd
import time
import numpy as np

df_unequal_refs = pd.read_csv("refs_different_merged.csv")

df_unequal_refs.head()

def replace_eleven_refs(row):
    ref = row['ref']
    if ref == "robbie hart":
        return "robert hart"
    elif ref == "steve lodge":
        return "stephen lodge"
    elif ref == "john e martin":
        return "john martin"
    elif ref == "t lloyd":
        return "john lloyd"
    elif ref == "andy d' urso":
        return "andy d'urso"
    elif ref == "michael jones":
        return "mike jones"
    else:
        return ref
    
def replace_prem_refs(row):
    ref = row['ref2']
    if ref == "gary lewis":
        return "ray lewis"
    elif ref == "paul ward":
        return "tony ward"
    elif ref == "michael riley":
        return "mike riley"
    elif ref == "robert madley":
        return "bobby madley"
    else:
        return ref

initial = df_unequal_refs
initial_refs_different = initial.query("ref != ref2")
print("Initial refs different: {}".format(initial_refs_different.shape[0]))

round_one = initial_refs_different
round_one["ref"] = initial.apply(replace_eleven_refs,axis=1)
round_one["ref2"] = initial.apply(replace_prem_refs,axis=1)
one_refs_different = round_one.query("ref != ref2")
print("Round one refs different: {}".format(one_refs_different.shape[0]))
one_refs_different.to_csv("round_one_refs_difference.csv",index=False)

round_two = one_refs_different.copy()
round_two["ref"] = one_refs_different.apply(replace_eleven_refs,axis=1)
round_two["ref2"] = one_refs_different.apply(replace_prem_refs,axis=1)
two_refs_different = round_two.query("ref != ref2")
print("Round two refs different: {}".format(two_refs_different.shape[0]))
two_refs_different.to_csv("round_two_refs_different.csv",index=False)

def decide_correct_ref(row):
    ref1 = row['ref']
    ref2 = row['ref2']
    if ref1 != ref2:
        if ref1 == "robbie hart":
            return "robert hart"
        elif ref1 == "steve lodge":
            return "stephen lodge"
        elif ref1 == "john e martin":
            return "john martin"
        elif ref1 == "t lloyd":
            return "john lloyd"
        elif ref1 == "andy d' urso":
            return "andy d'urso"
        elif ref1 == "michael jones":
            return "mike jones"
        else:
            return ref1
    else:
        return ref1   

merged = pd.read_csv("initial_merge_all_games_matching.csv")
merged['final_ref'] = merged.apply(decide_correct_ref,axis=1)
sorted(merged['final_ref'].unique())

merged.head()

merged.to_csv("fixed_referees_step5.csv",index=False)

import ast

def check_score(row):
    prem_score = row['score2'].split("-")
    prem_home = int(prem_score[0])
    prem_away = int(prem_score[1])
    eleven_home = len(ast.literal_eval(row['h_scorers']))
    eleven_away = len(ast.literal_eval(row['a_scorers']))
    if prem_home == eleven_home and prem_away == eleven_away:
        return True
    else:
#         print("Prem: {}, eleven: {}-{}".format(row['score2'],eleven_home,eleven_away))
#         print("")
        return False

merged['score_equal'] = merged.apply(check_score,axis=1)

merged['score_equal'].value_counts()

scores_not_equal = merged.query("score_equal == False")
scores_not_equal.to_csv("scores_not_equal.csv",index=False)
