==> Data Cleaning and Fixing/Combine 11v11 and Prem/step6-get-missing-goalscorers.ipynb <==
import pandas as pd
import time
import numpy as np
import requests
from bs4 import BeautifulSoup as soup

example_url = "https://www.11v11.com/matches/manchester-city-v-brighton-and-hove-albion-09-may-2018-357538/"
response = requests.get(example_url,headers={'User-Agent': 'Custom'})
page = soup(response.content, "html.parser")

def get_all_goal_scorers(page):
    goalsBlocks = page.findAll('div',{'class':'goals'})
    if len(goalsBlocks) == 0:
        return {'home':[],'away':[]}
    goalsBlock = goalsBlocks[0]
    homeGoalsBlock = goalsBlock.findAll('div',{'class':'home'})[0]
    all_home_goal_blocks = homeGoalsBlock.findAll('tr')
    all_home_goalscorers = []
    for i in range(len(all_home_goal_blocks)):
        home_goal_block = all_home_goal_blocks[i]
        home_goalscorer = home_goal_block.findAll('td')[0].text
        if home_goalscorer[0] == " ":
            home_goalscorer = home_goalscorer[1:]
        formatted_name = home_goalscorer.lower().replace(" ","-")
        all_home_goalscorers.append(formatted_name)

    awayGoalsBlock = goalsBlock.findAll('div',{'class':'away'})[0]
    all_away_goal_blocks = awayGoalsBlock.findAll('tr')
    all_away_goalscorers = []
    for i in range(len(all_away_goal_blocks)):
        away_goal_block = all_away_goal_blocks[i]
        away_goalscorer = away_goal_block.findAll('td')[0].text
        if away_goalscorer[0] == " ":
            away_goalscorer = away_goalscorer[1:]
        formatted_name = away_goalscorer.lower().replace(" ","-")
        all_away_goalscorers.append(formatted_name)
        
    all_goalscorers = {'home':all_home_goalscorers,'away':all_away_goalscorers}
    return all_goalscorers

all_goalscorers = get_all_goal_scorers(page)
print(all_goalscorers)

def get_lineups_data(page):
    lineupData = page.findAll('div',{'class':'lineup'})

    homeData = lineupData[0].findAll('div',{'class':'home'})
    homeLineupBlocks = homeData[0].findAll('div',{'class':'player flagged'})
    homePlayers = [x.a['href'].split("/")[2] for x in homeLineupBlocks]

    awayData = lineupData[0].findAll('div',{'class':'away'})
    awayLineupBlocks = awayData[0].findAll('div',{'class':'player flagged'})
    awayPlayers = [x.a['href'].split("/")[2] for x in awayLineupBlocks]
    
    return {'home':homePlayers,'away':awayPlayers}

lineups = get_lineups_data(page)
print(lineups)

def get_goalscorers_ids(page):
    goalscorers = get_all_goal_scorers(page)
    lineups = get_lineups_data(page)
    home_ids = []
    home_goalscorers = goalscorers.get("home")
    home_players = lineups.get("home")
    for goalscorer in home_goalscorers:
        for player in home_players:
            if goalscorer in player:
                home_ids.append(player.split("-")[-1])
    
    away_ids = []
    away_goalscorers = goalscorers.get("away")
    away_players = lineups.get("away")
    for goalscorer in away_goalscorers:
        for player in away_players:
            if goalscorer in player:
                away_ids.append(player.split("-")[-1])
                
    goalscorers_ids = {'home':home_ids,'away':away_ids}
    return goalscorers_ids

goalscorers_ids = get_goalscorers_ids(page)
print(goalscorers_ids)

missing_goalscorers_data = pd.read_csv("scores_not_equal.csv")
eleven_ids = missing_goalscorers_data['11id'].tolist()
urls = ["https://www.11v11.com/matches/" + str(int(x)) for x in eleven_ids]
urls[0:10]

def follow_link_get_goalscorers(url):
    response = requests.get(url,headers={'User-Agent': 'Custom'})
    page = soup(response.content, "html.parser")
    goalscorers = get_goalscorers_ids(page)
    return goalscorers

def get_home_goalscorers(row):
    elevenid = str(int(row['11id']))
    url = "https://www.11v11.com/matches/" + elevenid
    goalscorers = follow_link_get_goalscorers(url)
    home_scorers = goalscorers.get("home")
    print(elevenid)
    return home_scorers

def get_away_goalscorers(row):
    elevenid = str(int(row['11id']))
    url = "https://www.11v11.com/matches/" + elevenid
    goalscorers = follow_link_get_goalscorers(url)
    away_scorers = goalscorers.get("away")
    print(elevenid)
    return away_scorers

missing_goalscorers_data['home_scorers'] = missing_goalscorers_data.apply(get_home_goalscorers,axis=1)
missing_goalscorers_data['away_scorers'] = missing_goalscorers_data.apply(get_away_goalscorers,axis=1)

missing_goalscorers_data.to_csv("goalscorers_added_fixed_step6_2.csv",index=False)

missing_goalscorers_data.head()
