==> Data Cleaning and Fixing/Combine 11v11 and Prem/step7-merge-missing-goalscorers.ipynb <==
import pandas as pd
import numpy as np
import time

df_new_goalscorers = pd.read_csv("goalscorers_added_fixed_step6_2.csv")

eleven_home_goalscorers = {}
eleven_away_goalscorers = {}
for i in range(len(df_new_goalscorers)):
    row = df_new_goalscorers.iloc[i]
    eleven_home_goalscorers[row['11id']] = row['home_scorers']
    eleven_away_goalscorers[row['11id']] = row['away_scorers']

def replace_home_scorers(row):
    h_scorers = row['h_scorers']
    eleven_id = row['11id']
    try:
        goalscorers = eleven_home_goalscorers.get(eleven_id)
        if goalscorers != None:
            return goalscorers
        else:
            return h_scorers
    except Exception:
        return h_scorers
    
def replace_away_scorers(row):
    a_scorers = row['a_scorers']
    eleven_id = row['11id']
    try:
        goalscorers = eleven_away_goalscorers.get(eleven_id)
        if goalscorers != None:
            return goalscorers
        else:
            return a_scorers
    except Exception:
        return a_scorers
    
df_old = pd.read_csv("fixed_referees_step5.csv")
df_new = df_old.copy()
df_new['h_scorers'] = df_new.apply(replace_home_scorers,axis=1)
df_new['a_scorers'] = df_new.apply(replace_away_scorers,axis=1)
df_new.tail(10)

df_new.to_csv("fixed_goalscorers_step7.csv",index=False)
