==> Data Cleaning and Fixing/Combine 11v11 and Prem/step8-clean-and-validate.ipynb <==
import pandas as pd
import numpy as np
import time

df = pd.read_csv("fixed_goalscorers_step7.csv")

df.info()

def check_columns_agree(row,col1,col2):
    val1 = row[col1]
    val2 = row[col2]
    return (val1 == val2)

df['home_equal'] = df.apply(check_columns_agree,args=("home","home2"),axis=1)
df['home_equal'].value_counts()

df['away_equal'] = df.apply(check_columns_agree,args=("away","away2"),axis=1)
df['away_equal'].value_counts()

df['season_equal'] = df.apply(check_columns_agree,args=("season","season2"),axis=1)
df['season_equal'].value_counts()

df['date_equal'] = df.apply(check_columns_agree,args=("date","date2"),axis=1)
df['date_equal'].value_counts()

dates_disagree = df.query("date_equal == False")
dates_disagree

df['date_final'] = df['date']

df.head()

df_cols_dropped = df.drop(['refs_equal','home2','away2','ref2','ref','season2','date2','date','home_equal','away_equal','season_equal','date_equal'],axis=1)

df_cols_dropped.head()

df_cols_dropped.columns = ['premid','score','11id','season','home','away','stadium','h_scorers','a_scorers','h_players','a_players','ref','date']

df_cols_dropped.head()

df_final = df_cols_dropped[['premid','11id','season','date','home','score','away','ref','stadium','h_scorers','a_scorers','h_players','a_players']]

df_final.head()

df_final.to_csv("data_collected_final_version.csv",index=False)
